# Conflicts:
#	Application/Reporting/SerialLetter/Frontend.php

+++ Application/Reporting/SerialLetter/Frontend.php
@@ -2082,10 +2082,12 @@ class Frontend extends Extension implements IFrontendInterface
                 $tblRelationshipCompanyList = Relationship::useService()->getCompanyRelationshipAllByPerson($tblPerson);
                 if ($tblRelationshipCompanyList) {
                     foreach ($tblRelationshipCompanyList as $tblRelationshipCompany) {
-                        if ($Item['CompanyRelationship'] === '') {
-                            $Item['CompanyRelationship'] = $tblRelationshipCompany->getTblType()->getName();
-                        } else {
-                            $Item['CompanyRelationship'] .= ', '.$tblRelationshipCompany->getTblType()->getName();
+                        if ($tblRelationshipCompany->getServiceTblCompany()) {
+                            if ($Item['CompanyRelationship'] === '') {
+                                $Item['CompanyRelationship'] = $tblRelationshipCompany->getTblType()->getName();
+                            } else {
+                                $Item['CompanyRelationship'] .= ', '.$tblRelationshipCompany->getTblType()->getName();
+                            }
                         }
                     }
                 }
@@ -2533,14 +2535,15 @@ class Frontend extends Extension implements IFrontendInterface
         if ($Check === null) {
             $tblAddressAllByPerson = SerialLetter::useService()->getAddressPersonAllByPerson($tblSerialLetter, $tblPerson);
             if ($tblAddressAllByPerson) {
+                $FilterCategory = ( $tblFilterCategory ? $tblFilterCategory : null );
                 foreach ($tblAddressAllByPerson as $tblAddressPerson) {
                     if ($tblAddressPerson->getServiceTblPerson()
                         && $tblAddressPerson->getServiceTblPersonToAddress()
-                        && $tblAddressPerson->getServiceTblToPerson()
+                        && $tblAddressPerson->getServiceTblToPerson($FilterCategory)
                     ) {
                         $Global->POST['Check']
                         [$tblAddressPerson->getServiceTblPerson()->getId()]
-                        [$tblAddressPerson->getServiceTblToPerson()->getId()]
+                        [$tblAddressPerson->getServiceTblToPerson($FilterCategory)->getId()]
                         ['Address'] = 1;
 
 //                        $Global->POST['Check']
@@ -2624,28 +2627,31 @@ class Frontend extends Extension implements IFrontendInterface
                             /** @var TblToCompany $tblToCompany */
                             foreach ($tblAddressToPersonList as $tblToCompany) {
                                 if ($tblToCompany->getTblAddress()) {
-
-                                    if ($tblToCompany->getServiceTblCompany()->getId() === $tblRelationship->getServiceTblCompany()->getId()) {
-                                        $RelationShip = $tblType->getName();
-                                        $subDataList[] = array(
-                                            'Salutation'   => $tblPerson->getSalutation(),
-                                            'Person'       => $tblPerson->getFullName(),
-                                            'Relationship' => $RelationShip,
-                                            'Address'      => new CheckBox('Check['.$tblPerson->getId().']['.$tblToCompany->getId().'][Address]',
-                                                '&nbsp; '.$tblToCompany->getTblAddress()->getGuiString(), 1)
-                                        );
+                                    if ($tblToCompany->getServiceTblCompany() && $tblRelationship->getServiceTblCompany()) {
+                                        if ($tblToCompany->getServiceTblCompany()->getId() === $tblRelationship->getServiceTblCompany()->getId()) {
+                                            $RelationShip = $tblType->getName();
+                                            $subDataList[] = array(
+                                                'Salutation'   => $tblPerson->getSalutation(),
+                                                'Person'       => $tblPerson->getFullName(),
+                                                'Relationship' => $RelationShip,
+                                                'Address'      => new CheckBox('Check['.$tblPerson->getId().']['.$tblToCompany->getId().'][Address]',
+                                                    '&nbsp; '.$tblToCompany->getTblAddress()->getGuiString(), 1)
+                                            );
+                                        }
                                     }
                                 } else {
-                                    if ($tblToCompany->getServiceTblCompany()->getId() === $tblRelationship->getServiceTblCompany()->getId()) {
-                                        $RelationShip = $tblType->getName();
-                                        /** @var TblToPerson $tblRelationship */
-                                        $subDataList[] = array(
-                                            'Salutation'   => $tblPerson->getSalutation(),
-                                            'Person'       => $tblPerson->getFullName(),
-                                            'Relationship' => $RelationShip,
-                                            'Address'      => new Warning(
-                                                new Exclamation().' Keine Adresse hinterlegt')
-                                        );
+                                    if ($tblToCompany->getServiceTblCompany() && $tblRelationship->getServiceTblCompany()) {
+                                        if ($tblToCompany->getServiceTblCompany()->getId() === $tblRelationship->getServiceTblCompany()->getId()) {
+                                            $RelationShip = $tblType->getName();
+                                            /** @var TblToPerson $tblRelationship */
+                                            $subDataList[] = array(
+                                                'Salutation'   => $tblPerson->getSalutation(),
+                                                'Person'       => $tblPerson->getFullName(),
+                                                'Relationship' => $RelationShip,
+                                                'Address'      => new Warning(
+                                                    new Exclamation().' Keine Adresse hinterlegt')
+                                            );
+                                        }
                                     }
                                 }
                             }
@@ -2915,14 +2921,16 @@ class Frontend extends Extension implements IFrontendInterface
                 $tblPersonList = $this->getSorter($tblPersonList)->sortObjectBy('LastFirstName', new StringGermanOrderSorter());
                 /** @var TblPerson $tblPerson */
                 foreach ($tblPersonList as $tblPerson) {
-                    $tblAddressPersonAllByPerson = SerialLetter::useService()->getAddressPersonAllByPerson($tblSerialLetter,
-                        $tblPerson);
-                    if ($tblAddressPersonAllByPerson) {
-                        /** @var TblAddressPerson $tblAddressPerson */
-                        foreach ($tblAddressPersonAllByPerson as $tblAddressPerson) {
-                            // clean data if missing address
-                            if (!$tblAddressPerson->getServiceTblToPerson()) {
-                                SerialLetter::useService()->destroySerialAddressPerson($tblAddressPerson);
+                    if (!$tblFilterCategory) {
+                        $tblAddressPersonAllByPerson = SerialLetter::useService()->getAddressPersonAllByPerson($tblSerialLetter,
+                            $tblPerson);
+                        if ($tblAddressPersonAllByPerson) {
+                            /** @var TblAddressPerson $tblAddressPerson */
+                            foreach ($tblAddressPersonAllByPerson as $tblAddressPerson) {
+                                // clean data if missing address
+                                if (!$tblAddressPerson->getServiceTblToPerson()) {
+                                    SerialLetter::useService()->destroySerialAddressPerson($tblAddressPerson);
+                                }
                             }
                         }
                     }
@@ -2936,7 +2944,7 @@ class Frontend extends Extension implements IFrontendInterface
                         $tblFilterCategory = $tblSerialLetter->getFilterCategory();
                         array_walk($tblAddressPersonAllByPerson, function (TblAddressPerson $tblAddressPerson) use (&$AddressList, $tblPerson, $Id, $tblSerialLetter, $tblFilterCategory) {
                             if ($tblFilterCategory && TblFilterCategory::IDENTIFIER_COMPANY_GROUP == $tblFilterCategory->getName()) {
-                                if (( $serviceTblCompanyToAddress = $tblAddressPerson->getServiceTblToPerson() )) {
+                                if (( $serviceTblCompanyToAddress = $tblAddressPerson->getServiceTblToPerson($tblFilterCategory) )) {
 //                                        $tblAddress = $serviceTblCompanyToAddress->getTblAddress();
                                     $AddressList[$tblAddressPerson->getId()]['Person'] =
                                         $tblPerson->getLastFirstName();
